+++ extensions/web/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2025-09-03
+
+- Updated `Dockerfile` to copy renamed `scripts/` directory instead of `tools/`
+
 ## 2025-08-12
 
 - Initial web extension implementation